Integrate with message-panel-service.
Falls back to the npm module when the service is not installed.

Ideally we'd have an npm module that handles the boilerplate of
registering for the service and using a reasonable fallback for
when the service is not available. I'm not sure how to do
this with the current Atom API.